# Faster Xor filter